applied Gal comments
You can make the experiment figures larger: one panel in a row with text beside.
This will fill a whole column. Especially the sparse experiment (lower figures).
Also, you have a lot of small-font text that you may want to break.

Also, we could add the convergence theorem.
Please also make the margins inside the columns having a consistent size
(the distance between the frame of the boxes and the text).
Using a large margin makes the poster more enticing to read.